% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros.R
\name{insert_images}
\alias{insert_images}
\title{Image Processing of flow characteristics of rain-on-snow experiments}
\usage{
insert_images(raw_img_path)
}
\value{
Return matrix with applied threshold
}
\description{
Returns RDS data of the 2D plot of image inserted
}
\examples{
insert_images()

}
\author{
Ibrahim Bello, \email{ibb4u2006@yahoo.com}

Johanna Bloecher
}
