% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Image Processing of flow characteristics of rain-on-snow experiments}
\usage{
apply_threshold(raw_data)
}
\value{
Return matrix with applied threshold
}
\description{
Input data must be 3 in length (rows, columns, number of rgb arrays)
Returns RDS data of the 2D plot of image inserted
}
\examples{
## To process raw image
raw_image <- readJPEG('path to image')
## Apply threshold to the raw image data
apply_threshold(raw_image)

}
\author{
Ibrahim Bello, \email{ibb4u2006@yahoo.com}
}
