% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros.R
\name{insert_images}
\alias{insert_images}
\title{Read raw image
Apply thresholds}
\usage{
insert_images(raw_img_path)
}
\value{
Return matrix with applied threshold
}
\description{
Read raw image
Apply thresholds
}
